requirements.txt
fastapi
uvicorn[standard]
msal
requests
python-dotenv